+++ packages/etherplay-wallet-connector-fuel/src/types.ts
@@ -1,3 +1,0 @@
-import {FuelConnector} from 'fuels';
-
-export type UnderlyingFuelProvider = FuelConnector;